requirements.txt
aiogram>=3.0.0
python-dotenv
asyncio
openai
httpx
duckduckgo-search
